Porting CameraDisconnectTest for camera2/cameraPipe
(1) Parameterized CameraDisconnectTest to run both camera2/cameraPipe
implementation
(2) Move CameraDisconnectTest to the core_test_app in order to
test both the camera2 and cameraPipe implementations.
(3) Use ActivityRule to launch the test activity since the
ActivityScenario cannot resume from the paused state correctly.

In the camera disconnect test, test step is
(1) Start CameraXTestActivity
(2) Start Camera2Activity, and CameraXTestActivity will going to background
(3) Close Camera2Activity, and CameraXTestActivity resume
If we use the ActivityScenario for the test, the CameraXTestActivity may
not able to resume at the step (3).
The ActivityRule seems work fine on the above scenario.

Bug: 183704747
Test: manually run CameraDisconnectTest on lab devices
Change-Id: I5ea78a2e9cacfb7ef88d65883c6b4c45f602ae13