[cxcp-3a-bridge-3] Transform Result3A to FocusMeteringResult
This mirrors the logic from FocusMeteringControl in CamreaX.
(a) check if the af state is focused_locked
(b) check if the metering regions were applied by matching the capture
result with the appropriate capture request.

In camerapipe, the way we wire and invoke listeners, we already make
sure that we are using the result from the same or later request, so
we don't need any explicit check in the adapter, just checking for auto
focus state should suffice.

ref discussion -
https://docs.google.com/document/d/1v4r1hchb8fC9HfuapfH5mVQzawOG6FOYuJdGx4mlTKE/edit?usp=sharing

Test: ./gradlew :camera:camera-camera2-pipe-integration:test
      ./gradlew :camera:camera-camera2-pipe-integration:assemble

Change-Id: I67276e8c8e830c71be52c7faa87c57b3e6081702