Use nullable nothing for default parameters of nullable type
Original problem encountered with JS compiler, where default parameter substitution was using irConst(null) typed as nullable type parameter. Referencing type parameters outside of the function is not allowed, so it is now replaced with nullable nothing.

Apart from that, this change updates stability inference to check consts and composites to avoid inferring default parameters with nothing type as unstable.

Change-Id: I9103a6ee648deb817bf927eec9ae9ba376582464