Move Macrobench to snapshot_and_release
Fixes: 184088466
Test: ./gradlew bench:b-m:cC bench:b-m-j:cC
Test: ./gradlew bench:b-m:updateApi bench:b-m-j:updateApi
Test: ./gradlew dokkaKotlinDocs # manually inspected
Test: ./gradlew benchmark:integration-tests:macrobenchmark:assembleDebug compose:integration-tests:macrobenchmark:assembleDebug
Relnote: N/A

Start tracking APIs, and generating API docs

In the future, PerfettoCaptureWrapper should be a minimal API surface
area in the macro module for PerfettoRule (or other future infra) to
capture measurements, to minimize LIBRARY_GROUP API surface area. This
will also avoid PerfettoRule needing to duplicate API level checks
from PerfettoCaptureWrapper.

Note that we do now use LIBRARY_GROUP_PREFIX in very limited capacity,
for internal macrobenchmark/perfetto capture usage of internal
features. Ideally, we would label them as "test-only" to signify that
they're only to be used in integration tests, but as Macrobenchmark is
a test library, this isn't a useful distinction.

Change-Id: I165b23e3867a0a18a431983ecdaa1a616e03c190