Run BaseDaoTest with KSP
BaseDaoTest triggered a bug in XProcessing where we were over-extending
variance resolution to java sources, which is incorrect.

Unfortunately, KSP does not yet tell us if a Origin.CLASS was derived
from a kotlin file or other (java etc).
https://github.com/google/ksp/issues/375

For that, I've expanded KspFieldOrdering to also detect true origin via
reflection (and renamed it to KspClassFileUtility). I also added method
ordering there as order of methods might matter for Auto Value classes.
It adds a performance penalty but we'll eventually get rid of it when
https://github.com/google/ksp/issues/250 is fixed.
Note that method ordering does not check for descriptions so it is not
100% perfect.

Bug: 160322705
Test: BaseDaoTest, MethodSpecHelperTest, KspClassFileUtilityTest

Change-Id: Iddfed0f72c438987a7f68ded62e9653275685189