Adding documentation and sample code usage for AutoMigration.
This change also removes the @RestrictTo annotation from the AutoMigration related files, and has API specific changes in that now the user is expected to implement the AutoMigrationSpec interface, and the generated code uses the user defined class as the callback instance.

Test: Tests added to AutoMigrationProcessorTest.kt to check that the user provided class is a class, is static if it is a nested class, implements AutoMigrationSpec and has no default constructor args.
Bug: 180395129, 181655460
Relnote: Adding documentation and sample code usage for AutoMigration.
Change-Id: Iee07288d1a7ccfcd3e746c4d18ba5c6cae8c6d7f