Avoid work when intializing a NavBackStackEntry
A NavBackStackEntry serves two purposes: it is
a holder for the NavDestination and arguments
of an entry on the back stack and it is a stateful
LifecycleOwner, etc.

As such, initializing the NavBackStackEntry should
not do any work on its own. Instead, that work
should only be done when the NavBackStackEntry
is added to the NavController's back stack (i.e.,
the maxLifecycle becomes at least CREATED). Thus,
any access to the state (be it via ViewModels or
SavedStateHandle) should only be done once the
NavBackStackEntry's Lifecycle is CREATED, thus
preventing developers from accessing state that
is never actually part of the NavController.

This has no impact on the existing use cases for
NavBackStackEntry as they are currently only
accessible via NavController's methods (which, by
necessity, means that they are already on the
back stack). However, this change gives us the
flexibility to provide this to Navigators
prior to adding it to the back stack.

Test: updated and existing tests pass
Change-Id: Ibd36a79c88b9c5ff73cdaf4250286d5f448fd20f