Stop allocating an EdgeEffect to get the edge effect type.
For devices earlier than S, the only edge effect type available is
TYPE_GLOW, so don't allocate an EdgeEffect just to use it to
retrieve the edge effect type. Note that this allocation is only
temporary in the code until S attributes are locked in, but
we must move forward for 1p apps use of RecyclerView.

Test: StretchEdgeEffectTest
Bug: 184738332
Change-Id: I72ab75f9e1d0792c2c569719a4ebf71b8f5427c8