Update the documentation for a null providerInfo.
The result of calling the provider chooser is currently documented to
have a null providerInfo as part of the
ComplicationProviderChooserResult if the operation was cancelled.

However, the atual implementation returns a null value for
ComplicationProviderChooserResult and a null providerInfo is used when
the empty complication is selected.

The public API is behaving correctly, and the returned boolean value
is accurate, but the documentation was the only bit that was
inaccurate.

Test: end-to-end test of sample watch face
Change-Id: I8bfb38859e74c87bb70e61099c5ba8c45dd4caf1