Rewrite input injection for compatibility w/ Robolectric
Input is no longer injected from a coroutine that blocks the thread from
which it is ran (runBlocking). Instead, we now switch to the main thread
(or stay on the main thread if we were already there) and inject all
input events from there.

We no longer delay/sleep between the input events. It was never clear
that this was necessary in the first place, because handling of input
events only uses the timestamps from the events, which are correct. In
between each event we forward the clock to make sure that any clock
driven process can make the progression necessary to correctly handle
the input.

Fix: 186203688
Test: ./gradlew compose:ui:ui-test-junit4:test
Relnote: "Compose tests can now be run on Robolectric. The following
limitations have so far been identified:
* There is no native bitmap, so `ImageBitmap()` leads to a
  NullPointerException.
* There is no drawing, so `captureToImage()` will indefinitely await the
  next draw pass (i.e. it deadlocks).
* There is no font loaded, so any text will be measured incorrectly. All
  characters have a fixed height of 20px and width of 1px.
More limitations are expected to be identified in the future."

Change-Id: I284face60a4e5f6942b3d92cac7d47056d8174f3