Fix restoreBackStack() when saveBackStack() is pending
Any FragmentTransaction happens as separate stages:
1) Generate the set of transactions that need to be applied
2) Execute the transactions
3) Move the Fragments to their expected state

When calling saveBackStack() followed by restoreBackStack()
without any intervening time or specifically calls to
executePendingTransactions(), step 1 is applied to both
the saveBackStack() and restoreBackStack() commands without
step 2 and 3 actually being executed yet.

This means that saveBackStack() needs to capture enough
state while executing step 1 such that the
restoreBackStack() has enough information to generate
the proper set of FragmentTransactions. As such, the
previous approach of using an "execute runnable"
(which would only run in step 2) is not sufficient.

Instead, saveBackStack() now creates a deep copy of
the full BackStackRecord and immediately saves its
state as part of step 1, thus ensuring that it is already
available when step 1 for restoreBackStack() is available.

On the flip side, because step 2 and 3 haven't been
executed yet when restoreBackStack()'s step 1 executes,
the Fragments that would not have had their state saved
and been destroyed, meaning that BackStackState must
now consider the case where the Fragment still exists
prior to trying to find the saved state of the previously
destroyed fragment.

Test: new SaveRestoreBackStackTest test passes
BUG: 80029773
Change-Id: I147cc168d712d49f7533f51e56776506b3c7698c