Add CarHardwareManager classes.
Added CarHardwareManager, CarInfo, CarSensors and CarValue classes
and their supporting data classes.

Also added service to CarContext which enables apps to get an
implementation of the CarHardwareManager. If no implementation
was included via additional libraries, then an Exception will be
thrown if it is requeted.

Relnote: "Added androidx.car.app.hardware classes for access to car
    specific data such as fuel, battery and speed."
Bug: 188138054
Bug: 188137613
Test: ./gradlew :car:app:app:checkApi && ./gradlew bOS
Change-Id: Iff3c92ebca831557ce2e475c5294e2701f6e5a2e