Retain focusable from attributes in AppCompatEditText
This is a bugfix to retain focusable from attributes in
AppCompatEditText constructor.

The call to setKeyListener to implement emoji2 forced the view to
focusable true, even if the view specified focusable false in the xml.

```
<androidx.appcompat.widget.AppCompatEditText
    ...
    android:enabled="false"
    android:focusable="false" />
```

This returns appcompat to it's previous behavior, and the bug only
existed in appcompat-1.4.0-alpha01 with the introduction of emoji2.

Relnote: "Fix bug in AppCompatEditText which would cause views to be
focusable even when android:focusable=\"false\" was specified in the
xml"
Test: New unit tests
Bug: b/188049943

Change-Id: Ib94127fc4c6f7fcef5d9bc8652bd9c86bcdfe42b