Add support for CompactChip.
These are specific types of chip that we want to support to help developers and are a good test case for the API can/should evolve. Clearly the developer could do all of the work themselves using the basic chip shaped. But part of the goal with the Wear Compose library is to make for developers to build the types of components we want and also to be opinionated about how they do so.

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 186762723
Change-Id: Ic8e50fcfe4451669db2419a799722e6b33e76825