Fix refresh icon remaining after receding
Fixes: 186557635

When a SwipeRefreshLayout contains a NestedScrollView, the
nested scroll events were preventing the refresh icon from
completely disappearing. If the user completely retracted
the refresh icon and started scrolling, the icon had moved
off the screen, but it still had a shadow. When the icon
wasn't retracted completely, it would animate to the same
position and then scale to 0 and the shadow would no longer
be visible.

This CL causes the icon to be completely removed regardless
of whether or not it was previously retracted.

Test: manual, and ran all swiperefreshlayout tests
Change-Id: Icf29ab1c4ad3ac536c40cb28d955c58ec87e3a9d