Allow changing LifecycleOwner
Fixes: 189357118

Moving a Compose View between LifecycleOwners was not
updating the LifecycleOwner of the composition. This
CL turns the viewTreeOwner into a MutableState so that
it can be observed. It also keeps the ViewTreeOwner
callback if it isn't attached to the window. We assume
that the lifecycle owner hasn't changed until it is
attached to a window.

Test: new test

Change-Id: I221ded67d119611f5b20aea3ca36eeb56c2cbc21