Move ProGuard rules from navigation-common-ktx to common
NavArgsLazy was moved to navigation-common, but
the related proguard-rules.pro wasn't moved with it.

This meant that apps that didn't specifically include
the (now empty) navigation-common-ktx artifact wouldn't
get the correct set of ProGuard rules.

Test: tested in sample app
BUG: 190082521
Relnote: "ProGuard rules for `by navArgs()` now properly
apply when using `navigation-common` or any artifacts
that depend on it."

Change-Id: I688009632403c98afe728d856ce91cd056ef32a5