Fix DynamicIncludeGraphNavigator handling of entries
A Navigator must always defer to NavigatorState to
create NavBackStackEntry instances, rather than calling
create() itself, thus ensuring that arguments are correctly
passed through to the final navigation graph.

This also updates the APIs to match other Navigators,
correctly deferring to the superclass implementation
for the Navigator v1 APIs.

Relnote: "Fixed an issue with `<include-dynamic>` where
arguments passed to the graph would not be correctly passed
to the dynamically included graph."
Test: tested in the sample app
BUG: 190786048

Change-Id: I3e11572a629c34aaee290fb1255f258ebf632d75