Make InitialPagingSource extend DataSource
The placeholder value for pagingSource provided to the initial value of
PagedList for LivePagedList and RxPagedList uses InitialPagingSource,
which causes Paging to throw an IllegalStateException when
PagedList.dataSource is accessed. This check was initially added to
prevent users from mistakenly accessing .dataSource when building
PagedList with a PagingSource, but breaks existing paging2 behavior
that migrating users may depend on.

This CL moves InitialPagingSource to be based on DataSource instead,
since it is only used in specifically those compatibility paths.

Relnote: "Accessing PagedList.dataSource from the initial value produced
  by LivePagedList or RxPagedList will no longer incorrectly throw an
  IllegalStateException"
Test: ./gradlew paging:paging-common:test
Test: ./gradlew paging:paging-runtime:cC
Test: ./gradlew paging:paging-rxjava2:cC
Test: ./gradlew paging:paging-rxjava3:cC
Change-Id: I96707a477f347bd19e7aeccdb52eca39d637297a