Fix crash in Compose when Google Assistant is enabled
Root having zero bounds is a valid case when its layout params are WRAP_CONTENT (default implementation of
ComponentActivity.setContent) and there is no other
semantics child. But we were filtering out root node if it has zero bounds
therefore not crearing accessibility node info for it. Google Assistant depends on accessibility info in
some cases and was crashishing because of that.
So now we create node info for Compose root node always.

Test: new test + manually with Google Assistant on
Fixes: 187740619
Change-Id: I86d78b5217d0948bc84f0f94886d648637a57e5f