Fix for ScrollableState.animateScrollBy() not supporting repeatable/overshoot animations
Previously the logic was manually figuring out the animation end by comparing the current value with the target value. It is not correct for cases like repeatable or overshoot animations where we can reach the target value multiple times during the animation. In fact we should just use the available coroutine animation apis which handles that nicely.

Test: new test in ScrollableTest
Fixes: 189448873
Change-Id: Id0c10478d1855c61cfaa1ed9098cf138a4c01048