identity: Fix SessionTranscript assumption in software implementation.
The Identity Credential implementation for when running on an Android
device without hardware-backed Identity Credential had an assumption
that SessionTranscript was an array of size two. This assumption isn't
true in the most recent (and final) version of ISO 18013-5 since
SessionTranscript is now an array of size three. Fix this problem by
removing any assumptions about the structure of SessionTranscript.

Hardware-backed Identity Credential implementations are not affected,
this bug only existed in the software implementation in the Jetpack

Bug: None
Test: All unit tests pass.
RelNote: "Fix assumption about SessionTranscript in software-backed
Identity Credential implementation."

Change-Id: If61cd625c56762d149a9a70e95c0a4b67e7253e5