Allow the NavGraph to be changed in NavHost
Since NavGraph now implements its own equals based on the data in the
graph rather than the instance, we can stop always remembering the
NavGraph and allow it to be changed on recompose.

RelNote: "You can now make changes to the graph of a NavHost. Graphs
with the same startDestination and destinations in the graph will be
considered equal and will not clear the navController back stack."
Test: setSameGraph
Bug: 175392262

Change-Id: I0b8dbcea4186232c3280c4a43be11e4fafcc6ce3