On R and lower, don't support fling while dragging.
Fixes: 191400436

On R and lower, we can't use the EdgeEffect.isFinished() to
determine if the animation should have a fling() applied to it
because the animation may be running while the scroll is
occurring. There's no way to know how much animation is left
on R and lower, so we just animate it away with onRelease()
and then stop interacting with it.

This CL fixes both RecyclerView and NestedScrollView.

Test: manual testing and new tests for R and lower
Change-Id: I7081b477a2b3b2d439f363e3df20240eaa94617b