Redefine QueryResultAdapter to contain multiple RowAdapters
With the work on multimap return types a single row can represent multiple POJOs and thus requires multiple row adapters. This changes make it so that result adapters are compose of multiple row adapters which align better with the multimap return types.

This does mean cursor mismatch validation can no longer be done in the row adapter as a single adapter is unaware if other unused columns will actually be used or not, therefore cursor mismatch validation is moved to the query processor once the results adapter along with its row adapter have been created.

Bug: 191693863
Test: room-compiler:test
Change-Id: I320ee991fb7f567ca72eb5438cac9a92bf515583