Keep the sent PointerEvent when canceled
Fixes: 191758125

After receiving a canceled pointer stream, the stored current
event was being read as the one prior to the cancellation. That
meant that forEachGesture was not detecting that the pointers
were raised.

This CL keeps the "cancel" event so all pointers are detected
as raised when an ACTION_CANCEL is sent.

Test: manual and adjusted test
Change-Id: I5b5b5da23b9e0566dd0521784afb1079170fbf7a
(cherry picked from commit 316e31c67b5b95a83256b8be5739d8e7de28c8e3)