Device quirk: output distortion on Samsung devices
The symptom of this device quirk is that the output of one or many streams, including PRIV, JPEG and/or YUV, can have an extra 25% crop and the cropped image is stretched to fill the Surface, which results in a distorted output. The streams can also have an extra 25% double crop, in which case the stretched image will not be distorted, but the FOV is smaller than it should be. For more discussion, see b/190203334 and go/samsung-camera-distortion.

Previously, we have 2 separate fixes for this, one in camera-view and one in camera-core. This CL replaces the 2 existing fixes with a new fix in the core. The new fix forces a set of resolution that have been verified working on these devices.

- Add a SelectResolutionQuirk and a SelectResolution workaround.
- In SupportedSurfaceCombination, use the workaround to prioritize the selected resolution and increase the max resolution for preview if necessary.
- Remove SamsungPreviewTargetAspectRatioQuirk and PreviewOneThirdWiderQuirkTest.

Bug: 190203334
Test: manual test and ./gradlew bOS
Change-Id: I4ae2f6293eb37fe6a1710714a7714b5625b6a5ec